add 404

+++ src/views/DashboardView.vue
@@ -168,7 +168,12 @@
                           </tr>
                         </thead>
                         <tbody class="bg-white divide-y divide-gray-200 dark:bg-gray-800 dark:divide-gray-700">
-                          <tr class="hover:bg-gray-100 dark:hover:bg-gray-700" v-for="book in books">
+                          <tr v-if="books.length<=0">
+                            <td colspan="4">
+                              No data
+                            </td>
+                          </tr>
+                          <tr class="hover:bg-gray-100 dark:hover:bg-gray-700" v-else v-for="book in books">
 
                             <td class="py-4 px-6 text-sm font-medium text-gray-900 whitespace-nowrap dark:text-white">
                               {{ book.title }}</td>
@@ -400,27 +405,25 @@ export default {
           }
 
         } else {
-          
+
           this.$showSweetAlert('error', responseData.message);
         }
       } catch (error) {
         console.error('Login failed:', error);
-        
+
       }
     },
-    async delete_book(id){
-      
-     this.$confirmDelete('Delete Book ?','').
-     then(x =>
-      {
-        if (x)
-        {
-          this.delete_book_process(id)
-        }
-      });
-     
+    async delete_book(id) {
+
+      this.$confirmDelete('Delete Book ?', '').
+        then(x => {
+          if (x) {
+            this.delete_book_process(id)
+          }
+        });
+
     },
-    async delete_book_process(id){
+    async delete_book_process(id) {
       try {
 
         const response = await this.$axios.delete('/api/books/' + id, {
